Increase DescribeParameters batch size to 50
DescribeParameters' default batch size is 10, but the API supports a batch size
up to 50. Increasing it to 50 reduces the number of requests made to
Parameter Store.

This should help reducing the likelyhood of getting rate limited by AWS
(we are currently running into "ThrottlingException: Rate exceeded"
every time we deploy)