Allow dots in secret keys.
Dot is pretty common in java configuration (property) files. There is no
technical reason why dot should not be allowed. On the other hand,
service names should not be allowed to have dots.

What is more, refactor current key parsing a bit and move it to ssmstore
where it really belong (IMHO).