Add ListRaw store method
Which returns secret keys and values, but *not* metadata. Use it for the production-like commands — exec and export — which don't need the metadata.

Currently it's just a wrapper around List, which uses the AWS DescribeParameters API and thus has access to all of the metadata. Because that API is intended for CLI (not production) usage, it has very strict rate-limits. Separating-out the production usages with the new ListRaw API will allow us to use the AWS GetParametersByPath method instead, which only returns the keys and values, but has much higher rate-limits.