Print env var keys on exec with global `-v` flag
It's useful to know the exact keys that are picked up by Chamber as that
information is not otherwise available anywhere else. A global `verbose`
flag is added, and can be used by other commands to output useful
information.

Care must be taken not to expose any secret values, as `-v` flag may be
used in production to help trace origin of env vars.