add write --skip-unchanged flag (#175)
By default a write attempt results in a new history entry even when the value is unchanged. Under most use this is no big deal but I've run into cases where we update secrets through automation, for example syncing during transitions between systems, and the history is filled with a bunch of non-changes. This adds a simple option to ignore writing the secret if the value is unchanged.