Update AWS SDK to v1.16.26 (#181)
Update to the newest available AWS SDK which is v1.16.26. The main reason behind it is to add support for the `credential_process` in AWS SDK configuration.

This is the first time I've worked with go since the introduction of modules, so let me know if there's more to it than just bumping the version in `go.mod`

This resolves #180.