cmd: use x/sys/unix.Exec
We're not supposed to use the syscall package directly, and unix.Exec
does a better job of ensuring that the appropriate command is issued
for each build/OS combination.

I believe it was written this way at first because the unix package
did not have an "Exec" function, which was an oversight that was
corrected at least a year ago.