Optimise ListRaw for S3-KMS backend
We've started using `chamber exec` in some interactive CLI tools and noticed that it was quite slow in many cases. This is because the S3 KMS backend was making an unnecessary request for each secret, in addition to fetching all of the secrets in one go via the index file.

`ListRaw` was originally implented as a wrapper for `List`. This was unnecessarily slow, since `ListRaw` doesn't need any additional info about each secret (which requires the additional per-key request). Now, `ListRaw` doesn't call `List`; instead, it just pulls the data it needs straight from the index file.

This reduces a typical chamber invocation for one of our services from 3 seconds to 0.65 seconds.