Make Go 1.19 our API version
This brings with it a LOT of dependency upgrades, which is probably a
good thing overall. Whether or not it breaks tests is a different
question. Note that this also includes a change to the declared
package path for analytics-go from
"gopkg.in/segmentio/analytics-go.v3"
to "github.com/segmentio/analytics-go/v3".