store: fail faster if not running in EC2
If you are running `chamber` outside of EC2 without a region
configured, the binary will take 5 seconds to fail because the default
ec2 metadata API timeout is 5 seconds.

Instead, take a few milliseconds to determine whether
we are even running on EC2. We attempt to do this by
checking the contents of different files on the disk, e.g.
/sys/devices/virtual/dmi/id/board_asset_tag, and then if all of those
fail, attempting to see if anything is listening on the metadata host.

This can help prevent the 5 second failure timeout.