chore: Upgrade to aws-sdk-go-v2 (#481)
This is a major update to chamber's support for the S3, SSM, and Secrets
Manager store implementations. Every effort was made to preserve
functionality, but there is one gap.

The v2 SDK does not expose a retryer field for a minimum throttle delay,
so that argument is currently ignored when constructing new SSM stores.
Support for the delay will be addressed later.

The v2 SDK does not offer "iface" interfaces for the various clients, so
instead interfaces tailored to what chamber uses are defined. For
testing, these new interfaces are mocked, and mock types are generated
using github.com/matryer/moq. You don't need moq to use chamber or even
to build it, but only if you are developing chamber and make a change to
an API interface.

Also, old code in the SSM store implementation that allowed it to work
without IAM permissions for ssm:GetParametersByPath has been eliminated.
The permissions have been expected for a long time now.

Co-authored-by: Ryan McKern <344926+mckern@users.noreply.github.com>