chore!: Remove deprecated support for non-path SSM API (#503)
Version 2.0 of chamber deprecated use of the non-path-based API for the
SSM parameter store. This commit removes support for it completely.

Since the deprecation occurred so long ago, this changeset does not
preserve the ability to migrate to using the path-based API. Migration
may be performed with a 2.x version of chamber prior to upgrading to
a 3.x version.

BREAKING CHANGE: Any users or clients still using the non-path-based SSM
API will see errors working with the SSM parameter store.