feat: Add tag-on-create (#516)
Store implementations may support tagging secrets immediately after they
are created, as a convenience and a basis for future enforcement of
tagging. As usual, only the SSM store implementation supports the new
feature.

The README is updated with documentation about both this new feature and
the recently added tag commands.